Phase 3: split web/app.py into an app factory + blueprints
web/app.py (2118 lines, 41 routes) becomes a compatibility shim around:

- web/extensions.py — config + game/stats repositories + cleanup
  scheduler, initialized eagerly at import exactly as before (including
  load_dotenv before get_config)
- web/factory.py — create_app(): ProxyFix, CORS, cache headers,
  blueprint registration, app-wide SPA 404 fallback
- web/helpers.py — shared non-mutating serialization: serialize_state,
  compute_champions, format_end_game_results, battle export/parse, and
  a get_game_or_404 helper
- web/routes/core.py — /, /api/health, SPA fallback
- web/routes/game.py — battle lifecycle (state/start/judge/next/undo/
  end/tiebreak/export/upload/results)
- web/routes/spotify.py — OAuth token store + all /api/spotify/*
- web/routes/admin_stats.py — admin auth (origin check, rate limit) +
  /api/stats/*

The import surface is preserved: `from web.app import app, games` still
works (unit_tests.py), wsgi.py unchanged, `python web/app.py` still
runs the dev server.

Deliberate behavior normalizations (previously inconsistent per route):
- unknown session_id now returns 404 "Game not found" everywhere
  (several routes returned 400 "Invalid session ID")
- JSON bodies parsed with silent=True: missing/malformed bodies get a
  clean 400 instead of a 415/500
