+++ web/routes/__init__.py
@@ -1,0 +1,1 @@
+"""Route blueprints for the Hustle n' Tussle web API."""
